requirements.txt
requests
anthropic
python-dotenv
pandas
numpy
beautifulsoup4
websocket-client
schedule
flask
